resume_ai_matcher/requirements.txt
flask==3.0.0
scikit-learn==1.3.2
pdfplumber==0.10.3
nltk==3.8.1
numpy==1.26.2
google-generativeai==0.3.1
werkzeug==3.0.1
sentence-transformers
